feat(navbar): implement responsive mobile menu
